# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ad site API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/ads": {
            "get": {
                "tags": [
                    "Ads"
                ],
                "summary": "Get list of ads",
                "operationId": "App\\Http\\Controllers\\AdController::index",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "item sort",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "price"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "sort order",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "AdResponse",
                                    "description": "Ad response",
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Ad"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string"
                                        },
                                        "from": {
                                            "type": "integer"
                                        },
                                        "next_page_url": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "prev_page_url": {
                                            "type": "string"
                                        },
                                        "to": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ads/{id}": {
            "get": {
                "tags": [
                    "Ads"
                ],
                "summary": "Get ad",
                "operationId": "App\\Http\\Controllers\\AdController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ad id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Optional fields (created_at, description, images) comma separated",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ad"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/ads": {
            "post": {
                "tags": [
                    "Ads"
                ],
                "summary": "Create ad",
                "operationId": "App\\Http\\Controllers\\AdController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ad"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Ad": {
                "title": "Ad",
                "description": "Ad model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "link": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Image"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Ad"
                }
            },
            "Image": {
                "title": "Image",
                "description": "Image model",
                "properties": {
                    "link": {
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Image"
                }
            }
        }
    },
    "security": [
        []
    ]
}